Move user email address services to `email_addresses.py`
