+++ apps/users/services/__init__.py
@@ -1,2 +1,3 @@
 from .base import *
 from .email import *
+from .email_addresses import *
